test: mapped UI contracts tests
