+++ Makefile
@@ -233,14 +233,6 @@ test-user-configurator-used-as-collateral:
 test-rate-strategy:
 	make TEST_TARGET=rate-strategy.spec.ts test
 
-.PHONY: test-ui-pool-data-provider
-test-ui-pool-data-provider:
-	make TEST_TARGET=ui-pool-data-provider.spec.ts test
-
-.PHONY: test-ui-incentive-data-provider
-test-ui-incentive-data-provider:
-	make TEST_TARGET=ui-incentive-data-provider.spec.ts test
-
 .PHONY: test-reserve-configuration
 test-reserve-configuration:
 	make TEST_TARGET=reserve-configuration.spec.ts test
@@ -253,9 +245,9 @@ test-dynamic-configs-strategy:
 test-scenario:
 	make TEST_TARGET=scenario.spec.ts test
 
-.PHONY: test-wallet-balance-provider
-test-wallet-balance-provider:
-	make TEST_TARGET=wallet-balance-provider.spec.ts test
+.PHONY: test-ui-providers
+test-ui-providers:
+	make TEST_TARGET=_ui_providers.spec.ts test
 
 .PHONY: run
 run:
